- Template Matching   * Added correlation using FFT   * Tweaked API so that preprocessing of image can be performed   * Clarified how score can be used in example

+++ main/ip/src/boofcv/alg/misc/ImageStatistics.java
@@ -2196,8 +2196,8 @@ public class ImageStatistics {
 	 * @param img Input image.  Not modified.
 	 * @return Mean pixel intensity value
 	 */
-	public static double mean( GrayF32 img ) {
-		return sum(img)/(double)(img.width*img.height);
+	public static float mean( GrayF32 img ) {
+		return sum(img)/(float)(img.width*img.height);
 	}
 
 	/**
